requirements.txt
asgiref @ file:///C:/ci_311/asgiref_1676448296195/work
beautifulsoup4==4.12.2
Django @ file:///C:/ci_311/django_1676469589007/work
django-admin==2.0.2
django-excel-response2==3.0.5
django-livereload-server==0.4
django-rest-auth==0.9.5
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
excel-base==1.0.4
isoweek==1.3.3
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
screen==1.0.1
six==1.16.0
soupsieve==2.4.1
sqlparse @ file:///C:/ci_311/sqlparse_1676456541280/work
TimeConvert==3.0.13
tornado==6.3.2
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
tzlocal==5.0.1
xlwt==1.3.0
